Implement run_once
Prior to this patch, run_once was basically ignored. This patch adds
tracking on tasks that are tagged with run_once and will only allow it
to be invoked once. This means if the run_once task fails, all the other
hosts will continue to process their tasks like nothing happened. The
play itself will exit if the run_once task fails at the end. For tripleo
this means that run_once should not be used for any required tasks
within a play but can be used to make sure a task will only run once
time for a given play.

Change-Id: Iaa1278763d56429b2886371a18477e1f54680e19