Max percentage failure support
In this patch we update the new strategies so they support a max failure
percentage on a play. Which will be useful for the tripleo roles so we
can tolerate some % of failure (e.g. on computes) during a deployment.

Co-Authored-By: Alex Schultz <aschultz@redhat.com>
Co-Authored-By: Kevin Carter <kecarter@redhat.com>
Co-Authored-By: Emilien Macchi <emilien@redhat.com>

Change-Id: I30a930e4f7b8200ffa67c17b967db8a1fb60b4c0
(cherry picked from commit 87f9c9354884eace00bf7d42fbaa775055217e3f)