Molecule testing for TripleO Strategy Plugins
For both tripleo_free and tripleo_linear, add molecule tests that will
execute a simple task (for now) and test the max_fail_percentage value
that is defined in the Ansible inventory.

We simulate a failure on one compute node, which shouldn't make the
playbook failing since we tolerate 50% of failure (we have 6 nodes and
one failure).

Change-Id: I97a0171f9e7ba08f020ac7cf224a9ff9f6a34d01