Make throttle optional
Tasks did not have the concept of throttling prior to ansible 2.9 so if
we want to allow this to be re-used with older versions then we need to
skip the throttling code if task doesn't contain the throttle attribute.
This will help if we want to backport this for upgrades.

Change-Id: I01f6e3a69dec25ec69782c087d6b81ce3f33e6c0
(cherry picked from commit ee3c9c841e466c7fed7639bf3b5e19dbbe42abba)