Module tripleo_list_overclouds
When upgrading the undercloud it will be necessary to have some
ansible tasks to loop on tasks which operate on every existing stack.

This simple module will lookup existing heat stacks and return a list
of stacks which appear to be overcloud stacks. They are assumed to be
overcloud stacks if they have the output: AnsibleHostVarsMap.

Change-Id: Ib7dc2c968a5aaff40696f591caa2399c0fb5a7bc
Blueprint: nova-less-deploy