Preserve inode when updating /etc/hosts
Containers configured by tripleo bind-mount /etc/hosts directly,
which means any change to that file has to preserve the original
inode, otherwise the containers will get out of sync with the
host and will not see updates.

Change tripleo_host_entries to not depend on atomic semantics of
ansible, which changes inodes on update. Instead, perform a
non-atomic update to preserve inodes, and rely on the retry
logics of openstack containers to recover from unexpected
behaviour in case /etc/hosts is consumed while it is being updated.

Closes-Bug: #1882290
Change-Id: I34dd9121bbd650b79cb523e4dbed5949a0e7d52d