Add a new callback: tripleo_states
Summary: print out nodes which failed and were ignored

Within the context of max_fail_percentage, we want to improve UX by
displaying which nodes failed to be deployed or were unreachable, probably
because max_fail_percentage was set to a value which matches the play
results.

e.g. we tolerate 10% of compute node failures and two nodes failed
without blocking the whole deployment; the callback will display the
following message:

~~~~~~~~~~~~~~~~~~~ Summary Information ~~~~~~~~~~~~~~~~~~~~~~~~
~~~~~~~~~~~~~~~~~~Elapsed Time: 0:45:36.997 ~~~~~~~~~~~~~~~~~~~~
(...)
~~~~~~~ Number of nodes which did not deploy successfuly: 2 ~~~~
This or these node(s) failed to deploy: compute23, compute 45

In the future, this callback will be useful to improve state
informations that are displayed by Ansible when it's done running.

Change-Id: Ib337ea6c9c54a38cf11d021bce4edaf60e9928bd