Implement container_startup_config module
container_startup_config will replace a bunch of tasks that we did in
THT to generate the .json file for each container in
/var/lib/tripleo-config/container-startup-config.

It'll accelerate a bit the deployment by replacing tasks into a
single module, so we can generate the startup configs much faster.

Also tolerate empty configs in container_config_data, switching a
failure into a debug. If the config is empty, there is no need to run
the rest of the role tasks in tripleo_container_manage.

Note: we manage the SElinux context at the openstack-selinux:
https://github.com/redhat-openstack/openstack-selinux/commit/0b62

TODO (later): Manage idempotency and only create configs that need to be
updated.

Change-Id: I501f31b52db6e10826bd9a346d38e34f95ae8b75