Set short ANSIBLE_SSH_CONTROL_PATH_DIR for all ceph roles
We set ANSIBLE_SSH_CONTROL_PATH_DIR to deal with permissions
issues. Bug 1885917 was introduced because it was set to a
variable which can make the path too long for a Unix domain
socket. We can fix it by hard coding it to something short
which will not vary.

Change-Id: I57d5a8bb8b125727983df188968ae19366c3ebd4
Closes-Bug: #1885917
(cherry picked from commit ca165ec94018293dfbbbd37c44a2704b8144f203)