BaR: Deleted duplicated variables
The backup_and_restore has the same name for a
fact and a register variable, both of them named the same.
That variable is used later on a template (local.conf.j2), which
is using the register one.

This patch removed the register variable, as it's not used.

Change-Id: Id0487e4649ef1f6aaf2591e003746b385f592095
Closes-Bug: 1888404
(cherry picked from commit c384348f5586bb5fb47bf59b7bb9e0a61f8178bf)