Use ansible_user only if ANSIBLE_REMOTE_USER is unset
In a non-standalone deployment config-download is called
with the ANSIBLE_REMOTE_USER defined. This variable is
undefined when using standalone, which is what caused the
related bug. Unfortunately the fix for the related bug
introduced the bug this patch closes. This patch should
work for both non-standalone and standalone deployments.

Change-Id: Ie7234bf113204e8cf847257554b88b899b45d5ee
Related-Bug: #1886497
Closes-Bug: #1887708