tripleo_container_manage: start systemd services which would be stopped
If an operator would have disabled or stopped the service, we make sure
that it is again enabled and started when a deployment occurs.

We also add test coverage for it in molecule.

Change-Id: I97a6819574772835eb4a291b63bb05551fe4439a
(cherry picked from commit ddf494a16dd1f508fda3d46e2f7c8c7ab8d5d317)