Set ceph_ansible_limit to ansible_limit if ansible_limit defined
The 'openstack overcloud deploy ... --limit' command passes
the Ansible limit to the run of the config-download Ansible
but not to the run of ceph-ansible. To limit ceph-ansible
it's possible to override ceph_ansible_limit (an Ansible
variable) but, as per the bug this patch closes, this is
not possible via the deploy option of the tripleo client.
However, this patch makes it possible by having the variable
ceph_ansible_limit inherit the value of --limit but only
if it is passed. For backwards compatibility the inheritance
does not happen if ceph_ansible_limit is already defined.

Also, pass ceph_ansible_limit to the executions of the
machine unique UUID gather playbook and remote_tmp playbook.

Change-Id: Ibd1ca28592ba3dbbdd2dd129fe8319390f8435f3
Closes-Bug: 1888508
(cherry picked from commit 872233241bc2cdbd9a17cc61eff1d89f36e458d1)