Switch from handlers to tasks
Handlers normally don't run until the end of a play execution. Since we
are forcing a handler run in the middle of the tasks there is not a good
reason to use handlers for this case. Additionally we're not relying on
this handler anywhere else.  Because of this, let's switch to explicit
tasks that check changed results instead of handlers.  We've seen issues
with trying to use handlers in ansible when not using a linear strategy.

Change-Id: I4eb5ece7d5c1d25f5beacd9e21e20ae9f3d45c17
Closes-Bug: #1888853
(cherry picked from commit d057ad0b6a56110b49796bc2d6a331dd4ce9045f)