Add tripleo_ovs_upgrade module.
This module replace the code in [1] as we need to replicate this
sequence of actions in several places in the templates.

It first checks if it's a layered product ovs update which uses
rhosp-openvswitch.  If so, it makes sure that openvswitch is removed
without triggering any actions, like stopping the service.  Then we
install the rhos-openvswitch package which will take care of updating
everything without disturbance.

If the os is not using layer product, then we download the openvswitch
and update it without triggering the systemctl restart if present in
the postun scripts.

[1] https://opendev.org/openstack/tripleo-heat-templates/src/branch/master/deployment/tripleo-packages/tripleo-packages-baremetal-puppet.yaml#L234-L370

Change-Id: I793bdb8db1d34011cf64cf9942a811ab5efc00ed