Wait for node to be unlocked before provide
This change reduces the risk of provide having a lock timeout by
waiting for existing node locks to be released before starting the
provide.

Ansible based provide may not be affected by bug #1846791 because
power-down happens after the provide, not before. However waiting for
locks to be released is recommended practice, and doing it here may
improve reliability.

Change-Id: I5bced3b91e4fa3568185e2bbc85c0a000182394e
Closes-Bug: #1846791