Move first set of working mol jobs to CentOS-8
In order to port the molecule jobs from CentOS-7 to CentOS-8,
not all the jobs are working on CentOS-8 gathered via
https://review.opendev.org/#/c/747348/.
In order to make it easier, we have created a c8 based base job
and moved working jobs to c8.

Rest of the c7 based molecule jobs will be moved in a seperate
review.

Change-Id: I87ad037463eca1df41cc18aaccfb59ae2bf694f8
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>