Add tripleo_diff_exec module
This module takes a file input and checks to see if it has changed since
the last time checked and runs a command if it has changed. The module
will take a copy of the file being checked to be used for subsequent
executions if the file has changed.

Change-Id: Iea02a40639529ff9d80d3368f07ce81e6b1e911f
(cherry picked from commit 5702b7ba3dd4b9be7ba05332ff6ff4ba9b43d06e)