Only concatenate default_templates_dir when necessary
By default the default_templates_dir is appended to the
plan_environment file. This is fine unless the deployer
overrides the default plan_environment file with the -p
option. This patch adds jinja2 filters so that if the
plan_environment starts with a '/', then the default
templates directory is not appended. Otherwise the default
behavior is unchanged.

Change-Id: I5341da7fdea0a53b86295d74d68ef2013efb0ea0
Closes-Bug: #1893861
(cherry picked from commit 295555839c5e1d8e7e5ab1bb3fb39537d09da989)