Derive HCI reseverved host memory without average guest workload
Add new option to tripleo_derive_hci_parameters module so that
if both the average_guest_cpu_utilization_percentage and the
average_guest_memory_size_in_mb are not passed, then a simpler
calculation is performed which computes the required memory for
the Ceph OSDs.

It is preferable for the user to pass the expected workload but
if this value is unknown it is still better to at least reserve
the bare minimum in memory for the OSDs so that they do not get
shutdown from an OOM error.

Change-Id: I68ad3783d3e0bf47fd63ab79f6ea0b0655c751c1