Configure the default podman network
Instead of blanking out the file, we need to configure it. An empty
file causes the `podman network` to not function and prevents us from
using containers without --net=host.

While we don't use this in tripleo today, it is beneficial to not leave
podman in a broken state on the undercloud or standalone if we actually
wanted to use podman networking for an external container.  This change
won't remove the network if the interface has been configured and is in
use as to not break 'podman network'

It should be noted that the default interface was changed in later
versions from cni0 to cni-podman0 see podman 3755 (podman > 1.6).

Change-Id: Id3ba3a531b8f86d411c843683c94a9d632cbf217
Closes-Bug: #1889510
Closes-Bug: #1898120