[TRAIN-Only] Allow configuration of the default podman network
In newer versions we've switched out the default network configuration
to actually be configured. However in Train, folks may have already
deployed with the existing configuration so we can't switch the
defaults.  This change adds the ansible variable to allow configuration
when needed but by default will still blank the network configuration as
previously done.

It should be noted that the default interface was changed in later
versions from cni0 to cni-podman0 see podman 3755 (podman > 1.6).

This change also adds the login molecule jobs back in that were removed
in the initial docker to podman conversion due to the lack of network
configuration.

Change-Id: Id3ba3a531b8f86d411c843683c94a9d632cbf217
Closes-Bug: #1889510
Closes-Bug: #1898120