Avoid reboot for node update even when kernelargs changed
During update/upgrade/scale, already deployed node should
not be rebooted, if the kernel args are changed. Reboot of
the node should be planned by the user with migration and
downtime window. With this change, the kernel args will be
updated to the already deployed node, but it will not be
applied with reboot. For fresh nodes, where kernel args
entry is not found, the reboot will be applied.

In case, when the existing node is deployed without kernel
args but during an update the kernel args are added. At this
time a reboot is required, because there are other services
which depends on the kernel args like hugepages for OvS-DPDK
and iommu for SR-IOV.

It leaves with another edge case, when the node is deployed
with only SR-IOV and OvS-DPDK is added later or vice-versa,
the node will not be rebooted. As such kind of deployments
require network config changes, it requires more planning.
This scenario has to be handled on case-by-case basis.
Closes-Bug: #1895105
Change-Id: I25fbad4ef76bccc45a06f78751f7efb906758b90