Add release version package
Currently this is only used downstream, however there is a version
package that can allow us to ensure specific package requirements are
met.  We can likely add this to the upstream via a tripleo-version or
something to that efffect but currently this does not exist. So this
change adds a task to ensure that we have the latest version of this
version package at deploy time.

Change-Id: I5fd88c8e2fae811e854eb0dfb462fae8d1ad8f0d