Fix compatibility issue in tripleo callback
The default callback sets some compatibility options but does not
properly do it in both the ways that are used. The show_per_host_start
is a compatibility option but the default callback uses get_option(...)
which checks for the existance in _plugin_options. Since it is not
defined there, the callback fails hard and the function no longer works.
This change adds our own compatibility layer to ensure the compatibility
options are set in _plugin_options until we can get a fix upstream.

Change-Id: I476aee9863ad63ce6d7ccf9350dae8540a37f017
Closes-Bug: #1895571
(cherry picked from commit 7eca0c70e388cf07badb5e13e89fc7e1938b82dd)