Add support for network config template
This adds support to use jinja template for network config.
Also adds ones to be used with default undercloud and
standalone.

Though the templates of both undercloud and standalone are
almost the same, separate ones are kept as before.

If someone wants to override the whole config as supported
earlier it can be done by choosing a different template.

Change-Id: I16be7ee44af8f52c8ef0705e2ab692148d65bc28