BaR: Keep /etc/exports content
Currently the backup and restore role when is installing the nfs server,
the role uses a template to generate the /etc/exports file. This means
that if there already an nfs server installed with its configuration
the /etc/exports will be overwritten.

This patch keeps the configuration of the /etc/exports and appends
the "new" part.

Closes-Bug: #1897503
Closes BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1878480

Change-Id: I0b943047be06153504fabe895711cce789621ac5
(cherry picked from commit 6293be89a09256278d27a7af5e650154970f9786)