Add image id compare
If the image has been updated on disk, the podman_container isn't
rebuild to use the new container.  This change adds an image id check to
the image check to see if the image name and tag are the same but the id
of the image has changed. It should be noted that the podman_container
does not fetch the updated image as it simply checks if it exists. A
user needs to use podman_image with force: True prior to running
podman_container to hit this condition.

Change-Id: Icb718d1c1ea9b829a532c37844ffb3c8b5db5dbe
Related-Bug: #1895974