Compute dashboard_frontend_vip according to the haproxy frontend
When Ceph dashboard is deployed, the mgr plugin is enabled on the
storage network, and, assuming that network is the value assigned
to 'CephDashboardNetwork', we need to change the way the frontend
vip is computed both in the standard scenario and when composable
networks are provided via tht.
The purpose of this change is to have a set of tasks to properly
compute the 'dashboard_frontend_vip' variable using the hieradata
generated for each service.
When StorageNetwork is present, 'ceph_dashboard_vip' will be set
and assigned to the haproxy frontend (because that's the network
used to expose and access the dashboard components), but when the
'ceph_dashboard_vip' ansible variable is not present, we can rely
(as assumed before) on ctlplane network.
This change is required to fix the tls-everywhere integration and
it is still valid for standard deployments.

Change-Id: Ie25aec842ed344a7bb313f51bb25eb51203dd72f