BaR doc update + lftp missing package
There is a ReaR option to copy files to the backup with sftp this
"feature" needs the lftp package. This patch adds the lftp package
to be installed when ReaR is installed.

The documentation regarding backup-and-restore was updated, the name
of the role uses underscores and dash. This patch updates the name
of the role on the documentation and modified the underscores with
dashes.

Change-Id: I5435780a280e9a1eecacab488de1f0e5165f430e
(cherry picked from commit 31eec764dae66f556007eb4f4cbfbe943c0d9f0c)
(cherry picked from commit 11445ff9e59c062472a8f44376d3e1006fabcad6)