Network ports module
Ansible module to manage network ports for overcloud
nodes. The main module option 'instances' takes a list
of instances similar to the one created by the
tripleo_baremetal_expand_roles module.

Additionally the 'stack_name' must be specified. Tags
will be added to each port resource created, one
with a stack_name hint and the other with the hostname
and a third with the ironic node uuid. Tags are also added
to metalsmith managed vif ports.

The tags is used to filter for already existing instance
ports on re-run, i.e idempotency. The tags will also be
used by tripleo-ansible-inventory to create an ansible
inventory prior to having a stack created.

On re-run existing ports will be updated, if the definition
changed.

The parameter 'concurrency' controls the maximum threads
to use for parallell processing.

Depends-On: https://review.opendev.org/761845
Depends-On: https://review.opendev.org/760536
Partial-Implements: blueprint network-data-v2-ports
Change-Id: Ie2874190c869abb8f9372acb6a45e93557090b2c