Provision workflow managed/unmanaged node support
Adds boolean option 'managed' to the YAML defining
baremetal deployments. When an instance is defined with
'managed: false' it indicates that the server is already
deployed.

Also adds option 'management_ip' to the YAML defining
baremetal deployments. Any instnace with 'managed: false'
must either have a 'fixed_ip' defining the IP address
of the already provisioned server. Or it must have a
'management_ip' set.

Adds module: tripleo_unmanaged_populate_environment
which merges the 'managed: false' instance to the
baremetal environment.

Adding the support for 'managed' in the YAML used to
define a baremetal deployment provides a unified UX for
pre-provisioned and ironic/metalsmith provisioned
instances. It also keeps the interface to manage neutron
ports for composable networks identical for for
pre-provisioned and ironic/metalsmith instances.

Partial-Implements: blueprint network-data-v2-ports
Change-Id: I19c1028664ee30ee1162c02e6efc723ca8816b14