Add config_drive support to baremetal provision YAML
This change exposes the config_drive property added to
metalsmith_instances and documented here[1]

A meta_data key 'instance-type' is always set with the role name as
the value, to provide an equivalent to nova setting this to the name
of the flavor.

A merge strategy is implemented so that meta_data and cloud_config
defaults values are merged with instance value, with instance values
having precedence when both are specified.

[1] https://review.opendev.org/762578

Change-Id: Id43c9ef40a67a048b3228367b3c84cc373611f7c