New option to trigger paunch configuration on only one container.
Adding the container_name option enables one to create playbook like
this:

    - hosts: Compute
      gather_facts: true
      become: true
      tasks:
        - name: update one container
          paunch:
            config: "/var/lib/tripleo-config/container-startup-config/step_4"
            config_id:
             - 'tripleo_step4'
            container_name: ovn_controller
            action: apply
            container_cli: podman
            managed_by: tripleo-Compute
            config_overrides:
              image: undercloud-0.ctlplane.redhat.local:8787/openstack-ovn-controller:16.1_20201020.1_hotfix
            cleanup: True

In this example we deliver an specific change to all Compute node
without running the entire update process.

This is only syntaxic sugar as running with:

    config: "/var/lib/tripleo-config/container-startup-config/step_4/hashed-ovn_controller.json"

has the same effect.

Note that config_overrides doesn't have the same semantic when we use
the container_name option as we don't require the user to pass the
full structure where it should be (still in the context of offering
some more convenience to the user):

   { <container_name>: {<changed_key>: <change>, ...} }

Here we prepend the container_name to the structure so that the user
focus on the container parameter.

This file has been deleted after ussuri so this patch starts at
ussuri.

Change-Id: I70ccfad3ea96dff44641bf70d459adb7cccb21d5