[TRAIN-Only] Remove __init__ from base strategy
There is a loading issues when running under py2 that causes errors
because of the way the base strategy is loaded. This change moves the
variable initialization out of the base and puts it in the strategies
themselves instead. It's a tiny bit of duplication but prevents loading
issues when running under py2.

Change-Id: I668590662dfed029a258bb21cb9c1100cee76652
Closes-Bug: #1904770