networks_lower group_var replace role_networks_lower
The 'networks_lower' group_var carries a mapping of
network.name to network.name_lower for all networks.

The external_bridge interface on DVR compute nodes
that do not have the 'External' network associated
with the role still need to be able to lookup the
name_lower for the 'External' network.

The lookup via role_networks_lower fail's since the
'External' network is'nt associated.

Also if the 'External' network is associated with the
role set the address and routes.

For the mutliple_nic_vlans add address and routes and
the vlan member interface on the bridge.

Note, also remove duplicate name entry on the 'Tenant'
network section.

Depends-On: https://review.opendev.org/763301
Closes-Bug: #1904809
Change-Id: I19a011bfbbcdbacbe257625310490077755a8b70