Restore network_data net order nicX contract
With the Heat multi-nic example templates the
loop.index was used to set the nicX for each
interface, and ordered iteration was done over
all networks in network_data.yaml. The ctlplane
was always nic1, then nic2 was mapped to the
first network in the network_data.yaml definition.

When converting multi-nic templates to ansible
the ordering contract between network_data.yaml
and the nicX for each network was broken because
iteration happens only over networks associated
with the role.

This change restores the ordering contract in
the multi-nic templates by iterating over the
'networks_all' group_var which holds all enabled
networks in the order of apperance in
network_data.yaml.

Depends-On: https://review.opendev.org/763497
Closes-Bug: #1904894
Change-Id: I9d2767d7ce4f24645684fde6044c38a4b920dbb1