[kernel] Updating facts before disabling NICs
When we delete and recreate a stack multiple times, it can lead to
outdated facts and cause tripleo-kernel to disable nic1 when it
shouldn't. This happens when we have a bond/team on nic1 at some point
during the deployment.

Closes-Bug: #1906082

Signed-off-by: David Vallee Delisle <dvd@redhat.com>
Change-Id: I808daf58b606c717ab1bbae7d3d869d5baa67352
(cherry picked from commit 1166412e3bc0d6822ab32f5b405c5ab7593d9883)