Expand roles honstname to role mapping
Expand roles is extended to return a hostname
to role mapping for all instanced. this map is
used to set the 'triple_role=$ROLE_NAME' tag
on ports.

The intent is to use these this tag in
conjunction with the pre-existing tags and
neutron network port information to set up the
in the tripleo-common code utilized by
tripleo-ansible-inventory.

Change-Id: I992d7a3cd946e42039ca0819c51f104b1231171b