Fix ci issues
1) Drop tripleo-common constraint

The new pip dependency resolver doesn't work well with links. See pip
issue #8253.  Also we should require a release of tripleo-common for CI
to function. This is similar to what we do with tripleoclient. This
change also bumps the lower bound for tripleo-common to the latest for
master which is 13.0.0. This should be adjusted for backports.

2) Exclude reqs and tox files from modules job temporarly, will be
fixed in next patch

Change-Id: I1363d911a39e4072a706c61d83d113db6f85f71c
Closes-Bug: #1908079