Don't specify --os-cloud undercloud (ephemeral heat)
It is not necessary to specify --os-cloud undercloud when running
openstackclient commands as the stackrc file is expected to already be
sourced.

Also, by removing the --os-cloud argument it allows the playbooks to
work when using an ephemeral Heat instance with the same command,
instead of having to use a different command depending on what type of
Heat is used.

Implements: blueprint ephemeral-heat-overcloud
Change-Id: I1e8b5e58d4464900abc7f3234e7520c44d7cb0c0