Fix mtu reference in standalone template
The standalone documentation instructs to use the
'InterfaceLocalMtu' parameter to define the MTU
for the interface.

The 'local_mtu' variable used in the standalone template
get's the value of {{role.name}}LocalMtu.

This changes the standalone template to use the value of
the 'ctlplane_mtu', which in the depends on is set from
the 'InterfaceLocalMtu' parameter.

Depends-On: https://review.opendev.org/769709
Change-Id: I95dac5daa3176286fa5f83d27368aa2a54c32186