tripleo_frr: add BGP TTL security
FRR supports enforcing Generalized TTL Security Mechanism (GTSM) where
only neighbors that are the specified number of hops away will be
allowed to become neighbors.

This patch adds a new option to set the number of hops allowed,
defaulting to 1 for strict security out of the box. Setting value to
zero or less will disable GTSM.

Change-Id: I1166f22fef8e3f6b825343b4e2792ce9cfb10547